Split out InteractiveRiskScores component without panel wrapper
