+++ packages/styled-components/src/Components/RiskChart/index.ts
@@ -1,1 +1,2 @@
+export { InteractiveRiskChart } from './InteractiveRiskChart';
 export { default as RiskChart } from './RiskChart';
